ajustes no flyway

+++ src/main/resources/db/migration/V1__criando_estado.sql
@@ -1,5 +1,5 @@
 create table estado(
-  id int8 not null,
+  id serial not null,
   nome varchar(50),
   sigla varchar(2),
   constraint estado_pk primary key (id)
