+++ src/main/resources/db/migration/V2__criando_cidade.sql
@@ -1,5 +1,5 @@
 create table cidade(
-  id int8 not null,
+  id serial not null,
   nome varchar(100) not null,
   id_estado int8 not null,
