+++ src/main/resources/db/migration/V4__criando_pessoa.sql
@@ -1,5 +1,5 @@
 create table pessoa(
-	id int8 not null,
+	id serial not null,
 	nome varchar(100) not null,
 	telefone varchar(50),
 	constraint pessoa_pk primary key (id)
